코드에서 역할과 책임을 분리합니다
- 컨트롤러와 저장소의 역할을 분리합니다.
