+++ app/src/main/java/com/codesoom/assignment/controllers/TaskController.java
@@ -1,6 +1,7 @@
 package com.codesoom.assignment.controllers;
 
 import com.codesoom.assignment.models.Task;
+import com.codesoom.assignment.models.TaskStorage;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -8,34 +9,21 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @RestController
 @RequestMapping("/tasks")
 @CrossOrigin(origins = "http://localhost:3000", maxAge = 1800)
 public class TaskController {
-    private List<Task> tasks = new ArrayList<>();
-    private Long newId = 0L;
+    private final TaskStorage taskStorage = new TaskStorage();
 
     @GetMapping
     public List<Task> list() {
-        return tasks;
+        return taskStorage.list();
     }
 
     @PostMapping
     public Task create(@RequestBody Task task) {
-        if (task.getTitle().isBlank()) {
-            // TODO error...
-        }
-
-        task.setId(generateId());
-        tasks.add(task);
-        return task;
-    }
-
-    private Long generateId() {
-        newId += 1;
-        return newId;
+        return taskStorage.create(task);
     }
 }
